Server: Scripts are now loaded before items & NPCs.
Database:
Added missing data to Item::Info.
Added tasks' loading for items.
Updated tasks' loading for NPCs.
Moved Lua scripts to tasks folder.

World:
Added a method to query a task.

Player:
Added a method for determining if an item can be equipped.
Added a method to use an item (and equip).
Added a method to equip an item.
Added a method to unequip an item.
Possibly fixed the addItem() method.

Item:
Added missing data to Item::Info.
Added a constructor to avoid invalid values.
Added documentation to the class.
Added a method to set the position.
Added methods to determine what is the item.

Msg / Networking :
Fixed the structure of the MsgItem packet.
Added support for using item.
Fixed the swapping of MsgNpc packet.

+++ src/Security/Cryptography/blowfish.cpp
@@ -451,7 +451,7 @@ test_blowfish()
         swap_block(tmp);
         bf.encipher(tmp);
         swap_block(tmp);
-        for (size_t x = 0; success && x < Blowfish::BLOCK_SIZE; ++x)
+        for (int32_t x = 0; success && x < Blowfish::BLOCK_SIZE; ++x)
         {
             success = (cipher[i][x] == tmp[x]);
 
@@ -462,7 +462,7 @@ test_blowfish()
         swap_block(tmp);
         bf.decipher(tmp);
         swap_block(tmp);
-        for (size_t x = 0; success && x < Blowfish::BLOCK_SIZE; ++x)
+        for (int32_t x = 0; success && x < Blowfish::BLOCK_SIZE; ++x)
         {
             success = (clear[i][x] == tmp[x]);
